Create styling structure, add global variables, create simple welcome screen
