+++ Frontend/styles/global.css
@@ -1,0 +1,74 @@
+:root {
+    /* Color styles */
+    --tertiary: rgba(0, 228, 228, 1);
+    --secondary: rgba(30, 220, 131, 1);
+    --primary: rgba(199, 128, 255, 1);
+    --dark: rgba(23, 0, 53, 1);
+    --light-gray: rgba(238, 252, 245, 1);
+    --white: rgba(255, 255, 255, 1);
+    --gradient1: linear-gradient(153.43deg, #00E4E4 0%, #C780FF 83.33%);
+    --gradient2: linear-gradient(153.43deg, #1EDC83 0%, #00E4E4 83.33%);
+    --gradient3: linear-gradient(153.43deg, #C780FF 16.67%, #1EDC83 100%);
+    /* Text-size styles */
+    /* base size: plain (18px) */
+    --h3: 1.67rem;
+    --h1: 3.61rem;
+    --span: 1.11rem;
+    --span-lg: 1.67rem;
+    --plain: 1rem;
+
+    /* Borders */
+    --border-radius: 10px;
+
+    /* Glassmorphism */
+    --blur-bg: rgba(255, 255, 255, 0.1);
+    --blur-border: 1px solid var(--white);
+    --blur-box-sizing: border-box;
+    --blur-backdrop: blur(75px);
+}
+
+html,
+body {
+    padding: 0;
+    margin: 0;
+    font-family: -apple-system, Poppins, BlinkMacSystemFont, Segoe UI, Roboto,
+        Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,
+        sans-serif;
+    color: var(--dark);
+}
+
+* {
+    box-sizing: border-box;
+}
+
+footer {
+    width: 100%;
+    height: 100px;
+    border-top: 1px solid #eaeaea;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+footer img {
+    margin-left: 0.5rem;
+}
+
+footer a {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+a {
+    color: inherit;
+    text-decoration: none;
+}
+
+.blur {
+    background: var(--blur-bg);
+    border: var(--blur-border);
+    box-sizing: var(--blur-box-sizing);
+    backdrop-filter: var(--blur-backdrop);
+    -webkit-backdrop-filter: var(--blur-backdrop);
+}
