+++ Frontend/styles/styles.css
@@ -1,0 +1,1 @@
+@import './global.css';
